requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.2.0=pypi_0
aom=3.3.0=h0e60522_1
astunparse=1.6.3=pypi_0
blas=2.115=mkl
blas-devel=3.9.0=15_win64_mkl
blosc=1.21.1=h74325e0_3
brotlipy=0.7.0=py39hb82d6ee_1004
bzip2=1.0.8=h8ffe710_4
c-blosc2=2.2.0=hdf67494_0
ca-certificates=2022.6.15=h5b45459_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=5.2.0=pypi_0
certifi=2022.6.15=py39hcbf5309_0
cffi=1.15.1=py39h0878f49_0
cfitsio=4.1.0=h5a969a9_0
charls=2.3.4=h39d44d4_0
charset-normalizer=2.1.0=pyhd8ed1ab_0
click=8.1.3=py39hcbf5309_0
cloudpickle=2.1.0=pyhd8ed1ab_0
colorama=0.4.5=pyhd8ed1ab_0
cpuonly=2.0=0
cryptography=37.0.1=py39h21b164f_0
cytoolz=0.12.0=py39hb82d6ee_0
dask-core=2022.7.0=pyhd8ed1ab_0
flask=2.1.3=pyhd8ed1ab_0
flask-cors=3.0.10=pypi_0
flatbuffers=1.12=pypi_0
freetype=2.10.4=h546665d_1
fsspec=2022.5.0=pyhd8ed1ab_0
gast=0.4.0=pypi_0
giflib=5.2.1=h8d14728_2
google-auth=2.9.1=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.47.0=pypi_0
h5py=3.7.0=pypi_0
hdf5=1.12.1=nompi_h57737ce_104
idna=3.3=pyhd8ed1ab_0
imagecodecs=2022.2.22=py39h43fe4e9_6
imageio=2.19.3=pyhcf75d05_0
importlib-metadata=4.12.0=pypi_0
intel-openmp=2022.1.0=h57928b3_3787
itsdangerous=2.1.2=pyhd8ed1ab_0
jinja2=3.1.2=pyhd8ed1ab_1
jpeg=9e=h8ffe710_2
jxrlib=1.1=h8ffe710_2
keras=2.9.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
krb5=1.19.3=hc8ab02b_0
lcms2=2.12=h2a16943_0
lerc=3.0=h0e60522_0
libaec=1.0.6=h39d44d4_0
libavif=0.10.1=h8ffe710_0
libblas=3.9.0=15_win64_mkl
libbrotlicommon=1.0.9=h8ffe710_7
libbrotlidec=1.0.9=h8ffe710_7
libbrotlienc=1.0.9=h8ffe710_7
libcblas=3.9.0=15_win64_mkl
libclang=14.0.1=pypi_0
libcurl=7.83.1=h789b8ee_0
libdeflate=1.12=h8ffe710_0
libffi=3.4.2=h8ffe710_5
liblapack=3.9.0=15_win64_mkl
liblapacke=3.9.0=15_win64_mkl
libpng=1.6.37=h1d00b33_3
libssh2=1.10.0=h9a1e1f7_2
libtiff=4.4.0=h2ed3b44_1
libuv=1.44.2=h8ffe710_0
libwebp=1.2.3=h8ffe710_0
libwebp-base=1.2.3=h8ffe710_1
libxcb=1.13=hcd874cb_1004
libzlib=1.2.12=h8ffe710_2
libzopfli=1.0.3=h0e60522_0
locket=1.0.0=pyhd8ed1ab_0
lz4-c=1.9.3=h8ffe710_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.4.1=pypi_0
markupsafe=2.1.1=py39hb82d6ee_1
mkl=2022.1.0=h6a75c08_874
mkl-devel=2022.1.0=h57928b3_875
mkl-include=2022.1.0=h6a75c08_874
msys2-conda-epoch=20160418=1
networkx=2.8.5=pyhd8ed1ab_0
numpy=1.23.1=py39h1a62c8c_0
oauthlib=3.2.0=pypi_0
opencv-contrib-python=4.6.0.66=pypi_0
openjpeg=2.4.0=hb211442_1
openssl=3.0.5=h8ffe710_0
opt-einsum=3.3.0=pypi_0
packaging=21.3=pyhd8ed1ab_0
partd=1.2.0=pyhd8ed1ab_0
pillow=9.2.0=py39ha53f419_0
pip=22.1.2=pyhd8ed1ab_0
protobuf=3.19.4=pypi_0
pthread-stubs=0.4=hcd874cb_1001
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pysocks=1.7.1=py39hcbf5309_5
python=3.9.13=hcf16a7b_0_cpython
python_abi=3.9=2_cp39
pytorch=1.12.0=py3.9_cpu_0
pytorch-mutex=1.0=cpu
pywavelets=1.3.0=py39h5d4886f_1
pyyaml=6.0=py39hb82d6ee_4
requests=2.28.1=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-image=0.19.3=py39h2e25243_0
scipy=1.8.1=py39hc0c34ad_0
setuptools=63.2.0=py39hcbf5309_0
six=1.16.0=pypi_0
snappy=1.1.9=h82413e6_1
sqlite=3.39.1=h8ffe710_0
tbb=2021.5.0=h2d74725_1
tensorboard=2.9.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.9.1=pypi_0
tensorflow-estimator=2.9.0=pypi_0
tensorflow-io-gcs-filesystem=0.26.0=pypi_0
termcolor=1.1.0=pypi_0
tifffile=2022.5.4=pyhd8ed1ab_0
tk=8.6.12=h8ffe710_0
toolz=0.12.0=pyhd8ed1ab_0
torchaudio=0.12.0=py39_cpu
torchmetrics=0.9.3=pypi_0
torchvision=0.13.0=py39_cpu
typing_extensions=4.3.0=pyha770c72_0
tzdata=2022a=h191b570_0
ucrt=10.0.20348.0=h57928b3_0
urllib3=1.26.10=pyhd8ed1ab_0
vc=14.2=hb210afc_6
vs2015_runtime=14.29.30037=h902a5da_6
werkzeug=2.1.2=pypi_0
wheel=0.37.1=pyhd8ed1ab_0
win_inet_pton=1.1.0=py39hcbf5309_4
wrapt=1.14.1=pypi_0
xorg-libxau=1.0.9=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xz=5.2.5=h62dcd97_1
yaml=0.2.5=h8ffe710_2
zfp=0.5.5=h0e60522_8
zipp=3.8.1=pypi_0
zlib=1.2.12=h8ffe710_2
zlib-ng=2.0.6=h8ffe710_0
zstd=1.5.2=h6255e5f_2
